requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
chardet==3.0.4
discord.py==1.3.4
idna==2.10
multidict==4.7.6
websockets==8.1
yarl==1.5.1

certifi~=2020.6.20
pip~=20.2
setuptools~=49.2.1
python-dotenv~=0.14.0